+++ NEWS.md
@@ -1,3 +1,8 @@
+# Development
+
+* Fix bug in `getResultsUpset()` that failed when using the absolute value
+filters (`|<|`, `|>|`)
+
 # 0.23.1
 
 * The release tarball includes version 0.3.7 of the web app